feat: Refactor process vuex module.
