+++ src/store/modules/ADempiere/process/index.js
@@ -1,0 +1,18 @@
+
+import state from './state.js'
+import mutations from './mutations.js'
+import actions from './actions.js'
+import getters from './getters.js'
+
+/**
+ * Process Vuex Module
+ * @author Edwin Betancourt <EdwinBetanc0urt@outlook.com>
+ */
+const processControl = {
+  state,
+  mutations,
+  actions,
+  getters
+}
+
+export default processControl
