+++ src/store/modules/ADempiere/process/state.js
@@ -1,0 +1,21 @@
+/**
+ * Process State
+ * @author Edwin Betancourt <EdwinBetanc0urt@outlook.com>
+ */
+export default {
+  inExecution: [], // process not response from server
+  isVisibleDialog: false,
+  reportObject: {},
+  reportList: [],
+  metadata: {},
+  process: [], // process to run finish
+  sessionProcess: [],
+  notificationProcess: [],
+  inRequestMetadata: [],
+  reportViewList: [],
+  totalResponse: 0,
+  totalRequest: 0,
+  totalSelection: 0,
+  errorSelection: 0,
+  successSelection: 0
+}
